SIMParameterEstimation

Documentation for SIMParameterEstimation.

SIMParameterEstimation.Window

SIMParameterEstimation.separation_matrix

Statistics.mean!

Statistics.mean!

SIMParameterEstimation.Window — Type

Window(arr::AbstractArray)

Window of indices of the array arr, representing a Cartesian product of the contained index intervals.

Useful for computing the overlap of two signals during cross-correlation.

julia> signal = [0 0 0 0 0 0 0 0 0;
                 0 0 0 1 1 1 0 0 0;
                 0 0 1 1 1 1 1 0 0;
                 0 1 1 1 1 1 1 1 0;
                 0 1 1 1 1 1 1 1 0;
                 0 0 1 1 1 1 1 0 0;
                 0 0 0 1 1 1 0 0 0;
                 0 0 0 0 0 0 0 0 0];

julia> signal = OffsetArrays.Origin(-4,-4)(signal)
8×9 OffsetArray(::Matrix{Int64}, -4:3, -4:4) with eltype Int64 with indices -4:3×-4:4:
 0  0  0  0  0  0  0  0  0
 0  0  0  1  1  1  0  0  0
 0  0  1  1  1  1  1  0  0
 0  1  1  1  1  1  1  1  0
 0  1  1  1  1  1  1  1  0
 0  0  1  1  1  1  1  0  0
 0  0  0  1  1  1  0  0  0
 0  0  0  0  0  0  0  0  0

julia> w_signal = Window(signal)
Window{2}(CartesianIndices((OffsetArrays.IdOffsetRange(values=-4:3, indices=-4:3), OffsetArrays.IdOffsetRange(values=-4:4, indices=-4:4))))

julia> template = [0 1 0;
                   1 1 1;
                   0 1 0];

julia> template = OffsetArrays.centered(template)
3×3 OffsetArray(::Matrix{Int64}, -1:1, -1:1) with eltype Int64 with indices -1:1×-1:1:
 0  1  0
 1  1  1
 0  1  0

julia> w_template = Window(template)
Window{2}(CartesianIndices((OffsetArrays.IdOffsetRange(values=-1:1, indices=-1:1), OffsetArrays.IdOffsetRange(values=-1:1, indices=-1:1))))

julia> @assert length(w_template) == 9

julia> intersect(w_signal, w_template .+ CartesianIndex((3, 3)))
Window{2}(CartesianIndices((2:3, 2:4)))

SIMParameterEstimation.separation_matrix — Function

separation_matrix(ϕ::NTuple{3,<:Real}, μ::NTuple{3,<:Real}=(1, 1, 1))
separation_matrix(ϕ_0::Real,...) # Assume equidistant phases

Construct a separation matrix M with eltype(M) = Complex{T} for given phase shifts ϕ and modulations μ.

The matrix M is such that for acquisitions $D_1(\tilde{k})$, $D_2(\tilde{k})$ and $D_3(\tilde{k})$ the components are separated as

\[ \begin{pmatrix} C_{0}(\tilde{k}) \\ C_{-1}(\tilde{k}) \\ C_{+1}(\tilde{k}) \end{pmatrix} = \begin{pmatrix} H(\tilde{k})S(\tilde{k}) \\ H(\tilde{k})S(\tilde{k} - \tilde{\nu}) \\ H(\tilde{k})S(\tilde{k} + \tilde{\nu}) \end{pmatrix} = \begin{pmatrix} 1 & \tfrac{\mu_1}{2} e^{-i\phi_1} & \tfrac{\mu_1}{2} e^{i \phi_1} \\ 1 & \tfrac{\mu_2}{2} e^{-i\phi_2} & \tfrac{\mu_2}{2} e^{i \phi_2} \\ 1 & \tfrac{\mu_3}{2} e^{-i\phi_3} & \tfrac{\mu_3}{2} e^{i \phi_3} \end{pmatrix}^{-1} \begin{pmatrix} D_{1}(\tilde{k}) \\ D_{2}(\tilde{k}) \\ D_{3}(\tilde{k}) \end{pmatrix}\]

Statistics.mean! — Function

mean!(out::AbstractArray, A::AbstractArray, set::SummingSet)

Compute the mean of array A for each shift with the specified set that depends on the template for the cross-correlation.

Warning: When the template does not contain the origin (CartesianIndex(0,...)) in its axes, it may return NaNs wherever the shifted template does not overlap with the signal.

Statistics.mean! — Method

 mean!(out::AbstractArray, f::AbstractArray, inds::SummingSet)

Compute the mean of f for each of its index with the given indice set inds.

Examples